refactor: extract const
