+++ src/components/Community/Benefits.tsx
@@ -1,53 +1,9 @@
 import React from "react";
 import FlareComponent from "../Home/FlareComponent";
-import Connection from "./Logo/Connection";
-import Coupon from "./Logo/Coupon";
-import Internship from "./Logo/Internship";
-import Experience from "./Logo/Experience";
-import Idea from "./Logo/Idea";
-import Expert from "./Logo/Expert";
 import Container from "../Common/Container";
+import { benefits } from "@/constants/benefits";
 
 const Benefits = () => {
-  const benefits = [
-    {
-      title: "Connection",
-      description:
-        "Our partners provide valuable connections that can help broaden your professional network. By collaborating with us.",
-      icon: <Connection />,
-    },
-    {
-      title: "Coupon",
-      description:
-        "As part of our collaboration, you will receive exclusive coupons that offer special privileges to join the event or special discount.",
-      icon: <Coupon />,
-    },
-    {
-      title: "Internship",
-      description:
-        "We offer opportunities for internships with leading blockchain companies.",
-      icon: <Internship />,
-    },
-    {
-      title: "Experience",
-      description:
-        "Working with blockchain followers provides a unique experience in Community Engagement.",
-      icon: <Experience />,
-    },
-    {
-      title: "Idea to Impact",
-      description:
-        "Collaborating with us gives you the chance to create and develop new projects.",
-      icon: <Idea />,
-    },
-    {
-      title: "Expert Insight",
-      description:
-        "Benefit from the opportunity to receive valuable advice from industry experts and speakers.",
-      icon: <Expert />,
-    },
-  ];
-
   return (
     <div className="bg-gradient-to-b from-[#0F082C] via-black to-black py-16">
       <Container>
