--- src/components/Home/Sponsors.tsx
@@ -1,7 +1,0 @@
-export default function Sponsors() {
-    return (
-        <div>
-            SPONSORS
-        </div>
-    )
-}
