+++ src/constants/benefits.tsx
@@ -1,0 +1,47 @@
+import Connection from "@/components/Community/Logo/Connection";
+import Coupon from "@/components/Community/Logo/Coupon";
+import Internship from "@/components/Community/Logo/Internship";
+import Experience from "@/components/Community/Logo/Experience";
+import Idea from "@/components/Community/Logo/Idea";
+import Expert from "@/components/Community/Logo/Expert";
+
+const benefits = [
+  {
+    title: "Connection",
+    description:
+      "Our partners provide valuable connections that can help broaden your professional network. By collaborating with us.",
+    icon: <Connection />,
+  },
+  {
+    title: "Coupon",
+    description:
+      "As part of our collaboration, you will receive exclusive coupons that offer special privileges to join the event or special discount.",
+    icon: <Coupon />,
+  },
+  {
+    title: "Internship",
+    description:
+      "We offer opportunities for internships with leading blockchain companies.",
+    icon: <Internship />,
+  },
+  {
+    title: "Experience",
+    description:
+      "Working with blockchain followers provides a unique experience in Community Engagement.",
+    icon: <Experience />,
+  },
+  {
+    title: "Idea to Impact",
+    description:
+      "Collaborating with us gives you the chance to create and develop new projects.",
+    icon: <Idea />,
+  },
+  {
+    title: "Expert Insight",
+    description:
+      "Benefit from the opportunity to receive valuable advice from industry experts and speakers.",
+    icon: <Expert />,
+  },
+];
+
+export { benefits };
