+++ src/constants/partners.ts
@@ -1,0 +1,40 @@
+import SCBX from "@/assets/partners/scb-x.png";
+import BitkubExchange from "@/assets/partners/bitkub-exchange.png";
+import ETHPadThai from "@/assets/partners/ethpadthai.png";
+import CryptoMind from "@/assets/partners/cryptomind.png";
+import IC from "@/assets/partners/ic.png";
+import Velo from "@/assets/partners/velo.png";
+import BinanceTH from "@/assets/partners/binance-th.png";
+
+const partners = [
+  {
+    name: "SCBx",
+    image: SCBX,
+  },
+  {
+    name: "Bitkub Exchange",
+    image: BitkubExchange,
+  },
+  {
+    name: "ETHPadThai",
+    image: ETHPadThai,
+  },
+  {
+    name: "CryptoMind",
+    image: CryptoMind,
+  },
+  {
+    name: "IC",
+    image: IC,
+  },
+  {
+    name: "Velo",
+    image: Velo,
+  },
+  {
+    name: "Binance TH",
+    image: BinanceTH,
+  },
+];
+
+export { partners };
